// typescript-client/src/types/ConfigCommand.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { AttachProof } from "./AttachProof";
import type { ClientCapabilities } from "./ClientCapabilities";

export type ConfigCommand = { "SetConnectionOrphan": { allow_orphan_sessions: boolean, } } | { "ClaimSession": { session_cid: bigint, only_if_orphaned: boolean, } } | { "DisconnectOrphan": { session_cid: bigint | null, } } | { "ReleaseSession": { session_cid: bigint, } } | { "AttachSession": { session_cid: bigint, proof: AttachProof, } } | { "ReportFocus": { session_cid: bigint, peer_cid: bigint | null, focused: boolean, } } | { "DeclareCapabilities": { capabilities: ClientCapabilities, } };